Introduce common launch handler to Vanilla
This gives a better base for future PRs.
